+++ rulebot/src/main/kotlin/com/kyledahlin/rulebot/DiscordCache.kt
@@ -21,35 +21,37 @@ import discord4j.core.`object`.entity.Guild
 import discord4j.core.`object`.entity.Member
 import discord4j.core.`object`.entity.MessageChannel
 import discord4j.core.`object`.util.Snowflake
-import discord4j.core.spec.MessageCreateSpec
-import discord4j.core.spec.VoiceChannelJoinSpec
-import suspendChannel
-import suspendCreateMessage
-import suspendDelete
-import suspendGetMessageById
-import suspendVoiceState
 import java.util.*
 import javax.inject.Inject
 
 private const val CACHE_SIZE = 20
 
 /**
- * Hold the [CACHE_SIZE] latest messages and the corresponding [MessageChannel] that they belong
- * to.
+ * Hold information about things that do not persist between RuleBot runs (wrappers, members, guilds)
  */
 @RuleBotScope
-internal class DiscordCache @Inject constructor() {
-    private val _messages: Deque<DiscordWrapper> = ArrayDeque()
-    private val _botIds = mutableSetOf<Snowflake>()
+class DiscordCache @Inject constructor() {
 
-    internal fun add(event: RuleBotEvent, channel: MessageChannel, guild: Guild, member: Member) {
+    private val _messages: Deque<EventWrapper> = ArrayDeque()
+    private val _botIds = mutableSetOf<Snowflake>()
+    private val _guilds = mutableSetOf<GuildWrapper>()
+
+    fun createEventWrapperEntry(event: RuleBotEvent, channel: MessageChannel, guild: Guild, member: Member) {
         if (_messages.size > CACHE_SIZE) {
             _messages.pollFirst()
         }
-        _messages.add(EventMetadata(event, channel, guild, member))
+        _messages.add(EventWrapperImpl(event, channel, guild, member))
     }
 
-    fun getMetadataForEvent(event: RuleBotEvent): DiscordWrapper? {
+    fun createGuildWrapperEntry(guild: Guild) {
+        _guilds.add(GuildWrapperImpl(guild))
+    }
+
+    fun getGuildWrapper(snowflake: Snowflake): GuildWrapper? {
+        return _guilds.firstOrNull { it.id == snowflake }
+    }
+
+    fun getWrapperForEvent(event: RuleBotEvent): EventWrapper? {
         return _messages.firstOrNull { it.event === event }
     }
 
@@ -58,62 +60,6 @@ internal class DiscordCache @Inject constructor() {
     }
 
     fun getBotIds(): Set<Snowflake> = _botIds
-}
-
-interface DiscordWrapper {
-
-    val event: RuleBotEvent
-
-    suspend fun deleteMessage()
-
-    suspend fun sendMessage(message: String)
-
-    suspend fun sendMessage(withSpec: MessageCreateSpec.() -> Unit)
-
-    suspend fun getGuildOwnerId(): Snowflake?
-
-    suspend fun getGuildId(): Snowflake?
-
-    val isDm: Boolean
-
-    fun getRoleIds(): Set<Snowflake>
-
-    suspend fun joinVoiceChannel(with: VoiceChannelJoinSpec.() -> Unit)
-}
-
-private class EventMetadata(
-    ruleBotEvent: RuleBotEvent,
-    val channel: MessageChannel,
-    val guild: Guild?,
-    val member: Member
-) : DiscordWrapper {
-
-    override val event = ruleBotEvent
-
-    override suspend fun deleteMessage() {
-        channel.suspendGetMessageById(event.id)?.suspendDelete()
-    }
-
-    override suspend fun sendMessage(message: String) {
-        channel.suspendCreateMessage(message)
-    }
-
-    override suspend fun sendMessage(withSpec: MessageCreateSpec.() -> Unit) {
-        channel.suspendCreateMessage(withSpec)
-    }
-
-    override suspend fun getGuildOwnerId() = guild?.ownerId
-
-    override suspend fun getGuildId() = guild?.id
-
-    override fun getRoleIds() = member.roleIds
-
-    override val isDm = guild == null
-
-    override suspend fun joinVoiceChannel(with: VoiceChannelJoinSpec.() -> Unit) {
-        member.suspendVoiceState()?.suspendChannel()?.join(with)?.subscribe()
-    }
-
 }
 
 inline class AttachmentUrl(val url: String)
